Cleanup `common` package
