+++ pkg/operation/botanist/tunnel_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/gardener/gardener/pkg/client/kubernetes/fake"
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	"github.com/gardener/gardener/pkg/operation/botanist"
-	"github.com/gardener/gardener/pkg/operation/common"
 	"github.com/gardener/gardener/pkg/utils/test"
 )
 
@@ -55,7 +54,7 @@ var _ = Describe("Tunnel", func() {
 			ctx = context.Background()
 			cl = mockclient.NewMockClient(ctrl)
 			log = logr.Discard()
-			tunnelName = common.VPNTunnel
+			tunnelName = "vpn-shoot"
 			tunnelPod = corev1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Namespace: metav1.NamespaceSystem,
