+++ pkg/utils/miscellaneous_test.go
@@ -15,6 +15,8 @@
 package utils_test
 
 import (
+	"fmt"
+	"net"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -190,6 +192,113 @@ baz`, spaces)).To(Equal(`foo
 
 		It("should return map[string]string", func() {
 			Expect(InterfaceMapToStringMap(input)).Should(Equal(output))
+		})
+	})
+
+	Describe("#ComputeOffsetIP", func() {
+		Context("IPv4", func() {
+			It("should return a cluster IPv4 IP", func() {
+				_, subnet, _ := net.ParseCIDR("100.64.0.0/13")
+				result, err := ComputeOffsetIP(subnet, 10)
+
+				Expect(err).NotTo(HaveOccurred())
+
+				Expect(result).To(HaveLen(net.IPv4len))
+				Expect(result).To(Equal(net.ParseIP("100.64.0.10").To4()))
+			})
+
+			It("should return error if subnet nil is passed", func() {
+				result, err := ComputeOffsetIP(nil, 10)
+
+				Expect(err).To(HaveOccurred())
+				Expect(result).To(BeNil())
+			})
+
+			It("should return error if subnet is not big enough is passed", func() {
+				_, subnet, _ := net.ParseCIDR("100.64.0.0/32")
+				result, err := ComputeOffsetIP(subnet, 10)
+
+				Expect(err).To(HaveOccurred())
+				Expect(result).To(BeNil())
+			})
+
+			It("should return error if ip address is broadcast ip", func() {
+				_, subnet, _ := net.ParseCIDR("10.0.0.0/24")
+				result, err := ComputeOffsetIP(subnet, 255)
+
+				Expect(err).To(HaveOccurred())
+				Expect(result).To(BeNil())
+			})
+		})
+
+		Context("IPv6", func() {
+			It("should return a cluster IPv6 IP", func() {
+				_, subnet, _ := net.ParseCIDR("fc00::/8")
+				result, err := ComputeOffsetIP(subnet, 10)
+
+				Expect(err).NotTo(HaveOccurred())
+				Expect(result).To(HaveLen(net.IPv6len))
+				Expect(result).To(Equal(net.ParseIP("fc00::a")))
+			})
+
+			It("should return error if subnet nil is passed", func() {
+				result, err := ComputeOffsetIP(nil, 10)
+
+				Expect(err).To(HaveOccurred())
+				Expect(result).To(BeNil())
+			})
+
+			It("should return error if subnet is not big enough is passed", func() {
+				_, subnet, _ := net.ParseCIDR("fc00::/128")
+				result, err := ComputeOffsetIP(subnet, 10)
+
+				Expect(err).To(HaveOccurred())
+				Expect(result).To(BeNil())
+			})
+		})
+	})
+
+	Describe("#FilterEntriesByPrefix", func() {
+		var (
+			prefix  string
+			entries []string
+		)
+
+		BeforeEach(func() {
+			prefix = "role"
+			entries = []string{
+				"foo",
+				"bar",
+			}
+		})
+
+		It("should only return entries with prefix", func() {
+			expectedEntries := []string{
+				fmt.Sprintf("%s-%s", prefix, "foo"),
+				fmt.Sprintf("%s-%s", prefix, "bar"),
+			}
+
+			entries = append(entries, expectedEntries...)
+
+			result := FilterEntriesByPrefix(prefix, entries)
+			Expect(result).To(ContainElements(expectedEntries))
+		})
+
+		It("should return all entries", func() {
+			expectedEntries := []string{
+				fmt.Sprintf("%s-%s", prefix, "foo"),
+				fmt.Sprintf("%s-%s", prefix, "bar"),
+			}
+
+			entries = expectedEntries
+
+			result := FilterEntriesByPrefix(prefix, entries)
+			Expect(result).To(ContainElements(expectedEntries))
+		})
+
+		It("should return no entries", func() {
+			result := FilterEntriesByPrefix(prefix, entries)
+			Expect(result).To(BeEmpty())
 		})
 	})
 })
